Removed figure-markets from fees and added it to dex including the volume
